Chess: Board drawing
Not sure I understand why the drawing was done this way, but until I have a complete binary I want to create the exact same binary. I would like to extract some kind of bitmap blitting extension though to demonstrate the reason for using a DSL in an existing language.